requirements.txt
livereload
jupyter-book
matplotlib
numpy
sphinxcontrib-mermaid
sphinxcontrib-tikz
sphinx-exercise
sphinx-proof
ipython-sql
schemadisplay-magic
networkx
sympy
scipy
